# api/openapi.yaml
openapi: 3.0.3
info:
  title: FamilyQuest API
  version: 0.1.0
  description: REST API сервера управления детскими заданиями, сезонами и достижениями.
servers:
  - url: http://localhost:8080
    description: Local API server
...
paths:
...
  /api/v1/task/add:
    post:
      tags: [Tasks]
      summary: Создание базового задания
      operationId: addTask
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddTaskRequest"
      responses:
        "201":
          description: Задание создано
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessResponse"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/Task"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  responses:
    BadRequest:
      description: Некорректный запрос
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: Ошибка авторизации
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Forbidden:
      description: Недостаточно прав
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
  schemas:
    SuccessResponse:
      type: object
      required: [data]
      properties:
        data: {}
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          $ref: "#/components/schemas/ErrorData"
    ErrorData:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          example: VALIDATION_ERROR
        message:
          type: string
          example: Некорректный запрос
...
    AddTaskRequest:
      type: object
      required: [title, points]
      properties:
        title:
          type: string
        description:
          type: string
        points:
          type: integer
          format: int64
          minimum: 0
    Task:
      type: object
      required: [id, title, description, points]
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        description:
          type: string
        points:
          type: integer
          format: int64
...